change timestamps types to u64

+++ contracts/src/systems/config/interface.cairo
@@ -75,10 +75,7 @@ trait ICombatConfig<TContractState> {
 #[starknet::interface]
 trait INpcConfig<TContractState> {
     fn set_npc_config(
-        self: @TContractState,
-        world: IWorldDispatcher,
-        spawn_delay: u128,
-        pub_key: felt252
+        self: @TContractState, world: IWorldDispatcher, spawn_delay: u64, pub_key: felt252
     );
 }
